Replace Inputs::Base with Traits::Input
This makes loading easier as we no longer need to prioritise within a
single directory, plus traits seems like a more-logical place for the
shared behaviour.
