+++ lib/govuk_design_system_formbuilder/elements/inputs/text.rb
@@ -2,6 +2,12 @@ module GOVUKDesignSystemFormBuilder
   module Elements
     module Inputs
       class Text < Base
+        include Traits::Input
+        include Traits::Error
+        include Traits::Hint
+        include Traits::Label
+        include Traits::Supplemental
+
         def builder_method
           :text_field
         end
