+++ lib/govuk_design_system_formbuilder/elements/inputs/url.rb
@@ -2,6 +2,12 @@ module GOVUKDesignSystemFormBuilder
   module Elements
     module Inputs
       class URL < Base
+        include Traits::Input
+        include Traits::Error
+        include Traits::Hint
+        include Traits::Label
+        include Traits::Supplemental
+
         def builder_method
           :url_field
         end
